create new lca.lca_db module
