+++ sourmash/lca/__init__.py
@@ -4,3 +4,5 @@ from .command_summarize import summarize_main
 from .command_rankinfo import rankinfo_main
 from .command_gather import gather_main
 from .__main__ import main
+
+from .lca_db import LCA_Database
